refactor: :recycle: Added the modal to the navbar, so that it appears everywhere
